Moved UserProjectsFilter to UserInputs
