+++ src/specklepy/api/client.py
@@ -7,9 +7,9 @@ from specklepy.api.resources import (
     OtherUserResource,
     ProjectInviteResource,
     ProjectResource,
+    ServerResource,
     SubscriptionResource,
     VersionResource,
-    ServerResource,
     branch,
     commit,
     object,
