+++ src/specklepy/core/api/inputs/project_inputs.py
@@ -45,8 +45,3 @@ class ProjectUpdateRoleInput(BaseModel):
     userId: str
     projectId: str
     role: Optional[str]
-
-
-class UserProjectsFilter(BaseModel):
-    search: str
-    onlyWithRoles: Optional[Sequence[str]] = None
