+++ src/specklepy/core/api/resources/current/active_user_resource.py
@@ -4,8 +4,7 @@ from typing import List, Optional, overload
 from deprecated import deprecated
 from gql import gql
 
-from specklepy.core.api.inputs.project_inputs import UserProjectsFilter
-from specklepy.core.api.inputs.user_inputs import UserUpdateInput
+from specklepy.core.api.inputs.user_inputs import UserProjectsFilter, UserUpdateInput
 from specklepy.core.api.models import (
     ActivityCollection,
     PendingStreamCollaborator,
